move `run` handler into its own module
